+++ src/codecs/decode/market_overview.rs
@@ -11,6 +11,7 @@ pub fn market_overview_entry_from_proto(msg: &ProtoMarketOverview) -> MarketOver
         symbol_id: msg.symbol_id,
         symbol: msg.symbol.clone(),
         last_price: decode_price_ticks(msg.last_price_ticks, Some(msg.symbol.clone())),
+        index_price: decode_price_ticks(msg.index_price_ticks, Some(msg.symbol.clone())),
     }
 }
 
@@ -47,6 +48,7 @@ mod tests {
                 symbol_id: 2,
                 symbol: "ETH-USDT".into(),
                 last_price_ticks: 2_000_000,
+                index_price_ticks: 1_999_500,
                 ..Default::default()
             }],
             next_page_token: "tok".into(),
@@ -58,6 +60,10 @@ mod tests {
         assert_eq!(
             list.markets[0].last_price.as_ref().unwrap().as_ticks(),
             2_000_000
+        );
+        assert_eq!(
+            list.markets[0].index_price.as_ref().unwrap().as_ticks(),
+            1_999_500
         );
         assert_eq!(list.next_page_token, "tok");
     }